Refactor: rename 'queue_operations' to 'asynchronous_operations'
